updates based on jarvis integration
