+++ src/components/Cards/SuggestionCard/components/SuggestionCardDek.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+import { spacing } from '../../../../styles';
+
+const SuggestionCardDek = styled.div.attrs({
+  className: 'suggestion-card__dek',
+})`
+  margin-bottom: ${spacing.sm};
+`;
+
+export default SuggestionCardDek;
